# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AddBlacklistResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
  models.AuthorizeRequest:
    properties:
      password:
        example: 6a4a61f57bccf059abb82fc95589ebc428629326ab965390f25224e262455beb
        type: string
      username:
        example: vasyaPupkin
        maxLength: 25
        type: string
    required:
    - password
    - username
    type: object
  models.AuthorizeResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
      token:
        type: string
    type: object
...
  models.ConfirmFriendRequestResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
  models.DeleteFriendResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
  models.Error:
    properties:
      errorCode:
        type: integer
      errorDescription:
        type: string
    type: object
...
  models.SendFriendRequestResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API сервис пользователей для взаимодействия с сервисом авторизации
    и сервисом игровых комнат
  title: User Service API
  version: "1.0"
paths:
...
  /user/{id}/blackList/{username}:
    put:
      consumes:
      - application/json
      description: Позволяет добавить пользователя в черный список
      parameters:
      - description: Идентификатор пользователя, добавляющего в черный список
        in: path
        name: id
        required: true
        type: string
      - description: Имя пользователя, добавляемого в черный список
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AddBlacklistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AddBlacklistResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AddBlacklistResponse'
      summary: Добавление в черный список
      tags:
      - user
...
  /user/{id}/friends/{username}:
    delete:
      consumes:
      - application/json
      description: Позволяет удалить из друзей пользователя
      parameters:
      - description: Идентификатор пользователя, удаляющего друга
        in: path
        name: id
        required: true
        type: string
      - description: Имя пользователя, удаляемого из списка друзей
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteFriendResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DeleteFriendResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DeleteFriendResponse'
      summary: Удаление из друзей
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Позволяет отправить запрос дружбы от одного пользователя другому
      parameters:
      - description: Идентификатор пользователя, отправляющего запрос дружбы
        in: path
        name: id
        required: true
        type: string
      - description: Имя пользователя, получающего запрос дружбы
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SendFriendRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SendFriendRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SendFriendRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SendFriendRequestResponse'
      summary: Отправление запроса дружбы
      tags:
      - user
  /user/{id}/friends/{username}/{confirmation}:
    put:
      consumes:
      - application/json
      description: Позволяет ответить на запрос дружбы от одного пользователя другому
      parameters:
      - description: Идентификатор пользователя, принимающего запрос дружбы
        in: path
        name: id
        required: true
        type: string
      - description: Имя пользователя, отправившего запрос дружбы
        in: path
        name: username
        required: true
        type: string
      - description: Подтверждение запроса дружбы (YES или NO)
        in: path
        name: confirmation
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConfirmFriendRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ConfirmFriendRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ConfirmFriendRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ConfirmFriendRequestResponse'
      summary: Подтверждение/отклонение запроса дружбы
      tags:
      - user
...
  /user/authorize:
    post:
      consumes:
      - application/json
      description: Позволяет авторизовать пользователя в системе
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.AuthorizeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthorizeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AuthorizeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AuthorizeResponse'
      summary: Авторизация пользователя
      tags:
      - user
...
swagger: "2.0"